requirements.txt
Flask==3.1.0
opencv-python>=4.5.0
numpy>=1.21.0
tflite-runtime>=2.7.0
pyserial>=3.5
Werkzeug>=2.0.0
Jinja2>=3.0.0
